Update structure, put Form in separate componenet
